[update] layout docknize, some states, button etc

+++ include/ekg/draw/shape/shape.hpp
@@ -29,14 +29,14 @@
 namespace ekg::draw {
   void rect(
     const ekg::rect_t<float> &rect,
-    const ekg::rgba_t<float> &color,
+    const ekg::rgba_t<uint8_t> &color,
     ekg::pixel_thickness_t line_thicnkess,
     ekg::sampler_t &sampler
   );
 
   void rect(
     float x, float y, float w, float h,
-    const ekg::rgba_t<float> &color,
+    const ekg::rgba_t<uint8_t> &color,
     ekg::pixel_thickness_t line_thicnkess,
     ekg::sampler_t &sampler
   );
